refactor: unify weight type to i64 across MenuItem and Frontmatter
`MenuItem.weight` was `i32` while `Frontmatter.weight` was `Option<i64>`. TOML
integers parse as i64, so matching Frontmatter avoids cross-call-site casts and
keeps the source-of-truth type at every boundary.

+++ crates/kiln/src/config.rs
@@ -106,7 +106,7 @@ pub struct MenuItem {
 
     /// Sort order (ascending). Items without a weight default to 0.
     #[serde(default)]
-    pub weight: i32,
+    pub weight: i64,
 
     /// Whether this link points to an external site.
     #[serde(default)]
